colab_requirements.txt
# Baseline reqs
pandas>=2.2.2
numpy>=1.26.4
scikit-learn>=1.4.2
xgboost>=2.0.3
statsmodels>=0.14.2
pyyaml>=6.0.1

# LLM fine-tuning (Colab GPU)
transformers>=4.44.0
trl==0.9.6
accelerate>=0.33.0
datasets>=2.19.0
peft>=0.11.0
bitsandbytes>=0.43.1
